# docs/api.yaml
openapi: 3.1.0
info:
  title: Conex API
  version: 0.1.0
  description: OpenAPI specification of App.
servers:
  - url: https://conex.co.cr
...
paths:
...
  /{path}:
    get:
      tags:
        - Sites
      operationId: getSite
      summary: Render a public site
      parameters:
        - $ref: '#/components/parameters/SitePath'
      responses:
        '200':
          description: Public site HTML.
          content:
            text/html:
              schema:
                type: string
        default:
          $ref: '#/components/responses/Error'
...
  /api/sites:
    post:
      tags:
        - Sites
      operationId: createSite
      summary: Create a site
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
            example:
              path: go-fitness
              name: Go Fitness
              tags:
                - fitness
                - wellness
              html: <h1>Hello</h1>
      responses:
        '201':
          description: Site created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
              example:
                path: go-fitness
                public: false
                name: Go Fitness
                tags:
                  - fitness
                  - wellness
                url: https://conex.co.cr/go-fitness
                created_at: 1780310400
                last_modified: 1780310400
                clicks: 0
                sub: 1
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
    get:
      tags:
        - Sites
      operationId: listSites
      summary: List owned sites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sites owned by the authenticated account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'
              example:
                - path: go-fitness
                  public: false
                  name: Go Fitness
                  tags:
                    - fitness
                    - wellness
                  url: https://conex.co.cr/go-fitness
                  created_at: 1780310400
                  last_modified: 1780310400
                  clicks: 0
                  sub: 1
                - path: personal-blog
                  public: true
                  name: Personal Blog
                  tags:
                    - writing
                  url: https://conex.co.cr/personal-blog
                  created_at: 1780310400
                  last_modified: 1780310500
                  clicks: 12
                  sub: 1
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
  /api/sites/{path}:
    head:
      tags:
        - Sites
      operationId: sitePathAvailable
      summary: Check whether a site path is available
      parameters:
        - $ref: '#/components/parameters/SitePath'
      responses:
        '204':
          description: Site path is available.
        '409':
          description: Site path is already in use or reserved by another route.
        default:
          $ref: '#/components/responses/Error'
    get:
      tags:
        - Sites
      operationId: getOwnedSite
      summary: Get owned site metadata
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SitePath'
      responses:
        '200':
          description: Owned site metadata and stored HTML.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnedSite'
              example:
                path: go-fitness
                public: true
                name: Go Fitness
                tags:
                  - fitness
                  - wellness
                url: https://conex.co.cr/go-fitness
                created_at: 1780310400
                last_modified: 1780310500
                clicks: 12
                sub: 1
                html: <h1>Hello</h1>
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags:
        - Sites
      operationId: updateSite
      summary: Update a site
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SitePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteRequest'
            example:
              public: true
              name: Go Fitness Club
              tags:
                - fitness
                - health
              html: <h1>Hello again</h1>
      responses:
        '204':
          description: Site updated.
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags:
        - Sites
      operationId: deleteSite
      summary: Delete a site
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SitePath'
      responses:
        '204':
          description: Site deleted.
        '401':
          $ref: '#/components/responses/ReauthenticationRequired'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ReauthenticationRequired:
      description: The bearer access token is missing, invalid, or expired. The client should reauthenticate the user.
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    SitePath:
      name: path
      in: path
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 255
        pattern: '^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$'
...
  schemas:
...
    CreateSiteRequest:
      type: object
      additionalProperties: false
      required:
        - path
        - name
        - tags
        - html
      properties:
        path:
          type: string
          minLength: 3
          maxLength: 255
          pattern: '^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$'
        name:
          type: string
          minLength: 1
          maxLength: 255
        tags:
          type: array
          maxItems: 32
          items:
            type: string
            minLength: 1
            maxLength: 64
          uniqueItems: true
        html:
          type: string
          maxLength: 268435456
    UpdateSiteRequest:
      type: object
      additionalProperties: false
      properties:
        public:
          type: boolean
        name:
          type: string
          minLength: 1
          maxLength: 255
        tags:
          type: array
          maxItems: 32
          items:
            type: string
            minLength: 1
            maxLength: 64
          uniqueItems: true
        html:
          type: string
          maxLength: 268435456
    Site:
      type: object
      additionalProperties: false
      required:
        - sub
        - path
        - public
        - name
        - tags
        - url
        - created_at
        - last_modified
        - clicks
      properties:
        sub:
          type: integer
        path:
          type: string
        public:
          type: boolean
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        url:
          type: string
          description: Public site URL built from the configured base URL and site path.
        created_at:
          type: integer
          format: int64
          description: Site creation timestamp.
        last_modified:
          type: integer
          format: int64
          description: Site last-modified timestamp.
        clicks:
          type: integer
          format: int64
          minimum: 0
          description: Number of public site visits.
    OwnedSite:
      allOf:
        - $ref: '#/components/schemas/Site'
        - type: object
          additionalProperties: false
          required:
            - html
          properties:
            html:
              type: string
    Error:
      type: object
      description: Generic error payload shape when the server returns JSON errors.
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - msg
          properties:
            msg:
              type: string